// ui/src/generated-api-client/models/project-detail-request.ts
/* tslint:disable */
/* eslint-disable */
/**
 * UPOU TODO API
 * Task management API for UPOU students and faculty
 *
 * The version of the OpenAPI document: 1.0.0
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

// May contain unused imports in some cases
// @ts-ignore
import type { ViewEnum } from "./view-enum";

export interface ProjectDetailRequest {
  is_default?: boolean;
  title: string;
  view?: ViewEnum;
  order?: number;
}